GEOT-2894, FeatureVisitors on Long fields fail because of typo in CalcUtil
git-svn-id: http://svn.osgeo.org/geotools/trunk@34689 e5c1c795-43da-0310-a71f-fac65c449510

+++ modules/library/main/src/main/java/org/geotools/feature/visitor/CalcUtil.java
@@ -253,7 +253,7 @@ public class CalcUtil {
             return new Double(0);
         } else if (bestClass == Float.class) {
             return new Float(0);
-        } else if (bestClass == Float.class) {
+        } else if (bestClass == Long.class) {
             return new Long(0);
         } else if (bestClass == Integer.class) {
             return new Integer(0);
